Creating cuboid

+++ src/main/java/com/guciowons/simplecuboids/cuboid/Cuboid.java
@@ -2,16 +2,14 @@ package com.guciowons.simplecuboids.cuboid;
 
 import org.bukkit.entity.Player;
 
-import java.math.BigDecimal;
-
 public class Cuboid {
-    private BigDecimal locationX;
-    private BigDecimal locationY;
-    private BigDecimal locationZ;
+    private int locationX;
+    private int locationY;
+    private int locationZ;
     private Player player;
     private int size;
 
-    public Cuboid(BigDecimal locationX, BigDecimal locationY, BigDecimal locationZ, Player player, int size) {
+    public Cuboid(int locationX, int locationY, int locationZ, Player player, int size) {
         this.locationX = locationX;
         this.locationY = locationY;
         this.locationZ = locationZ;
@@ -19,15 +17,15 @@ public class Cuboid {
         this.size = size;
     }
 
-    public BigDecimal getLocationX() {
+    public int getLocationX() {
         return locationX;
     }
 
-    public BigDecimal getLocationY() {
+    public int getLocationY() {
         return locationY;
     }
 
-    public BigDecimal getLocationZ() {
+    public int getLocationZ() {
         return locationZ;
     }
